+++ WooCommerce/Classes/Extensions/UIDevice+Woo.swift
@@ -11,7 +11,7 @@ extension UIDevice {
     var modelIdentifier: String {
         var size = 0
         sysctlbyname("hw.machine", nil, &size, nil, 0)
-        var machine = [CChar](repeating: 0,  count: size)
+        var machine = [CChar](repeating: 0, count: size)
         sysctlbyname("hw.machine", &machine, &size, nil, 0)
 
         return String(cString: machine)
